ncc-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
alembic==1.13.3
asyncpg==0.31.0
psycopg2-binary==2.9.11
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
httpx==0.27.2
python-dotenv==1.0.1
websockets==13.1
bcrypt==4.2.0
packaging>=24.0

# ── Testing ────────────────────────────────────────────────────────────────────
pytest>=8.0
pytest-asyncio>=0.23
